fix: cents

+++ scripts/generaTabelle.sql
@@ -42,7 +42,7 @@ drop table if EXISTS Prodotto;
 
 create table Prodotto(
   id_prodotto int primary key AUTO_INCREMENT,
-  prezzo float(6, 3) not null,
+  prezzo float(6, 2) not null,
   nome varchar(255) not null,
   allergeni varchar(255) not null,
   ingredienti varchar(255) not null,
